Added the Elixir Projects
The first project deals with a dictionary, and inputs words to split them up and find every possible word in the word. The second project takes in a large string and redacts certain words from it, replacing the words with *'s based on it's length. The final project takes in an equation and finds it's integral based on initial conditions