Imported the Clojure Projects
The first project takes dates in, can return them in base 10 or word format, and can go a number of days into the future to display these dates. The second project takes in files and returns statistics like total sentences, words, etc. It can also cleanse the words of any punctuation, and counts the frequency of words